Changed the way tile uncovering works.
Now, the tile is uncovered when you release the A button instead of
press it. This is closer to the original game, and allows the smiley
face to display its neutral (frightened) face while the A button is
down, directly before a bomb may or may not be uncovered. Additionally,
a “tick” sound is played when the A button is pressed, and the original
“OK” or “cancel” sound is heard when the button is released.